backend/mapping-austrofascism/requirements.txt
Django==1.11.28
django-cors-headers==2.1.0
django-extensions==2.2.9
django-filter==1.1.0
djangorestframework==3.7.3
django-tinymce4-lite==1.6.0
djangorestframework-camel-case==0.2.0
django-location-field==2.0.6
Pillow==2.1.0
django-softhyphen==1.1.0
django-taggit==1.2.0
django-taggit-autosuggest==0.3.8
django-taggit-serializer==0.1.7
geopy==2.0.0
beautifulsoup4==4.9.3
